requirements.txt
amqp==5.3.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
annotated-types==0.7.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
argon2-cffi-bindings==21.2.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
argon2-cffi==23.1.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
asgiref==3.8.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
billiard==4.2.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
boto3==1.35.63 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
botocore==1.35.63 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
celery==5.4.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
certifi==2024.8.30 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
cffi==1.17.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
charset-normalizer==3.4.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
click-didyoumean==0.3.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
click-plugins==1.1.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
click-repl==0.3.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
click==8.1.7 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
colorama==0.4.6 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0" and (sys_platform == "win32" or platform_system == "Windows")
colorlog==6.9.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
contextlib2==21.6.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
cron-descriptor==1.4.5 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
cryptography==41.0.7 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
defusedxml==0.7.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-allauth==0.59.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-celery-beat==2.7.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-celery-results==2.3.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-cors-headers==4.6.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-debug-toolbar==4.4.6 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-extensions==3.2.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-ninja-extra==0.20.9 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-ninja-jwt==5.3.5 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-ninja==1.2.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-otp==1.5.4 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django-timezone-field==7.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
django==4.2.16 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
djangorestframework==3.15.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
feedparser==6.0.11 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
flower==2.0.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
h11==0.14.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
humanize==4.11.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
idna==3.10 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
injector==0.22.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
jmespath==1.0.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
kombu==5.4.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
oauthlib==3.2.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
prometheus-client==0.21.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
prompt-toolkit==3.0.48 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
psycopg2-binary==2.9.10 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pycparser==2.22 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pydantic-core==2.23.4 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pydantic==2.9.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pyjwt==2.9.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pyjwt[crypto]==2.9.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
python-crontab==3.2.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
python-dotenv==1.0.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
python3-openid==3.2.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
pytz==2023.4 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
requests-oauthlib==2.0.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
requests==2.32.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
s3transfer==0.10.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
sgmllib3k==1.0.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
six==1.16.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
sqlparse==0.5.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
tornado==6.4.1 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
typing-extensions==4.12.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
tzdata==2024.2 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
urllib3==2.2.3 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
uvicorn==0.30.6 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
vine==5.1.0 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
wcwidth==0.2.13 ; python_version >= "3.11.dev0" and python_version < "3.12.dev0"
